chore: move shared methods to base class
